requirements.txt
langchain
langchain-community
openai
chromadb
streamlit
python-dotenv
tiktoken
PyPDF2
pypdf
#All dependencies required
